+++ packages/frontend/src/app/AboutSection.tsx
