+++ packages/frontend/src/app/HeroSection.tsx
